refactor: LottoApp의 통계 로직을 LottoStatistics로 분리
